Rename m3u8->openm3u8 for the fork
