+++ LICENSE
@@ -1,8 +1,10 @@
-m3u8 is licensed under the MIT License:
+openm3u8 is licensed under the MIT License:
 
 The MIT License
 
 Copyright (c) 2012 globo.com webmedia@corp.globo.com
+
+Modifications Copyright (c) 2026 Wurl (solutions@wurl.com)
 
 Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
