+++ runtests
@@ -17,7 +17,7 @@ function stop_server {
 }
 
 function run {
-    py.test -vv --cov-report term-missing --cov m3u8 tests/
+    py.test -vv --cov-report term-missing --cov openm3u8 tests/
 }
 
 function main {
