+++ tests/test_variant_m3u8.py
@@ -1,10 +1,11 @@
 # Copyright 2014 Globo.com Player authors. All rights reserved.
+# Modifications Copyright (c) 2026 Wurl.
 # Use of this source code is governed by a MIT License
 # license that can be found in the LICENSE file.
 
 import playlists
 
-import m3u8
+import openm3u8 as m3u8
 
 
 def test_create_a_variant_m3u8_with_two_playlists():
